refactor models to single location
