+++ src/Device/DeviceFactory.cs
