+++ test/Exceptions/DeviceNotFoundExceptionTest.cs
